iex "repipe"
If input to iex results in a syntax error relating to |>, it checks if
the error is because the pipeline came at the beginning of the line. If
so, and if there is a prior value in iex's history, it uses that prior
value as the left-hand side of the pipeline. If not, it prints the error
as before.